pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "firebot"
version = "0.1.0"
description = "A modular, Python-first paper trading platform for ML-based strategy experimentation"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["trading", "paper-trading", "ml", "backtesting", "quantitative-finance"]
authors = [
    { name = "FireBot Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Intended Audience :: Science/Research",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "pydantic>=2.0",
    "polars>=0.20",
    "pandas>=2.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
ml = [
    "torch>=2.0",
]
parallel = [
    "ray>=2.9",
]
metrics = [
    "influxdb-client>=1.40",
    "prometheus-client>=0.19",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]
all = [
    "firebot[ml,parallel,metrics,dev]",
]

[project.urls]
Documentation = "https://github.com/firebot/firebot#readme"
Issues = "https://github.com/firebot/firebot/issues"
Source = "https://github.com/firebot/firebot"

[tool.hatch.build.targets.wheel]
packages = ["src/firebot"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
]

[tool.ruff.isort]
known-first-party = ["firebot"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
source = ["src/firebot"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
